Fix errors related to keymaps
`x-popup-menu` is sometimes passed a keymap or list of keymaps instead
of a list of panes, which `avy-menu` expects. If we receive the keymap
version of the argument instead, fall back to the original
`x-popup-menu` call. Ideally, we would be able to handle this case with
`avy-menu` as well, but falling back is a lot better than the current
implementation (which just fails with an error).

The error can be easily reproduced by clicking on one of the lighters in
the mode line.